feat: create key_crack module and moved pertinent methods to it as well as some to positionator
